[core] Rework state update methods and deprecate useGridApi + useGridState
