+++ packages/grid/_modules_/grid/hooks/core/useGridInitialization.ts
@@ -4,10 +4,10 @@ import type { GridComponentProps } from '../../GridComponentProps';
 import { useGridLoggerFactory } from './useGridLoggerFactory';
 import { useGridApiInitialization } from './useGridApiInitialization';
 import { useGridErrorHandler } from './useGridErrorHandler';
-import { useGridControlState } from './useGridControlState';
 import { useGridLocaleText } from './useGridLocaleText';
 import { useGridPreProcessing } from './preProcessing';
 import { useGridRowGroupsPreProcessing } from './rowGroupsPerProcessing';
+import { useGridStateInitialization } from './useGridStateInitialization';
 
 /**
  * Initialize the technical pieces of the DataGrid (logger, state, ...) that any DataGrid implementation needs
@@ -16,7 +16,7 @@ export const useGridInitialization = (apiRef: GridApiRef, props: GridComponentPr
   useGridLoggerFactory(apiRef, props);
   useGridApiInitialization(apiRef, props);
   useGridErrorHandler(apiRef, props);
-  useGridControlState(apiRef, props);
+  useGridStateInitialization(apiRef, props);
   useGridPreProcessing(apiRef);
   useGridRowGroupsPreProcessing(apiRef);
   useGridLocaleText(apiRef, props);
