+++ packages/grid/_modules_/grid/hooks/features/filter/useGridFilter.ts
@@ -36,7 +36,6 @@ const checkFilterModelValidity = (model: GridFilterModel) => {
  * @requires useGridColumns (state, method, event)
  * @requires useGridParamsApi (method)
  * @requires useGridRows (event)
- * @requires useGridControlState (method)
  */
 export const useGridFilter = (
   apiRef: GridApiRef,
