+++ packages/grid/_modules_/grid/hooks/features/pagination/useGridPage.ts
@@ -35,7 +35,6 @@ const applyValidPage = (paginationState: GridPaginationState): GridPaginationSta
 };
 
 /**
- * @requires useGridControlState (method)
  * @requires useGridPageSize (state, event)
  * @requires useGridFilter (state)
  */
