+++ packages/grid/_modules_/grid/hooks/features/selection/useGridSelection.ts
@@ -42,7 +42,6 @@ const useUtilityClasses = (ownerState: OwnerState) => {
 /**
  * @requires useGridRows (state, method)
  * @requires useGridParamsApi (method)
- * @requires useGridControlState (method)
  */
 export const useGridSelection = (
   apiRef: GridApiRef,
