+++ packages/grid/_modules_/grid/hooks/features/sorting/useGridSorting.ts
@@ -34,7 +34,6 @@ type GridFieldComparator = { field: string; comparator: GridComparatorFn };
 
 /**
  * @requires useGridRows (state, event)
- * @requires useGridControlState (method)
  * @requires useGridColumns (event)
  */
 export const useGridSorting = (
