+++ packages/grid/_modules_/grid/hooks/utils/useGridState.ts
@@ -1,43 +1,8 @@
-import * as React from 'react';
-import { GridEvents } from '../../models/events';
 import { GridApiRef } from '../../models/api/gridApiRef';
-import { GridState } from '../../models/gridState';
-import { useGridApi } from './useGridApi';
 
+/**
+ * @deprecated Use `apiRef.current.state`, `apiRef.current.setState` and `apiRef.current.forceUpdate` instead.
+ */
 export const useGridState = (
   apiRef: GridApiRef,
-): [GridState, (stateUpdaterFn: (oldState: GridState) => GridState) => boolean, () => void] => {
-  useGridApi(apiRef);
-
-  const forceUpdate = React.useCallback(
-    () => apiRef.current.forceUpdate(() => apiRef.current.state),
-    [apiRef],
-  );
-
-  const setGridState = React.useCallback(
-    (stateUpdaterFn: (oldState: GridState) => GridState) => {
-      const newState = stateUpdaterFn(apiRef.current.state);
-      if (apiRef.current.state === newState) {
-        return false;
-      }
-
-      const { ignoreSetState, postUpdate } =
-        apiRef.current.unstable_applyControlStateConstraint(newState);
-
-      if (!ignoreSetState) {
-        // We always assign it as we mutate rows for perf reason.
-        apiRef.current.state = newState;
-
-        if (apiRef.current.publishEvent) {
-          apiRef.current.publishEvent(GridEvents.stateChange, newState);
-        }
-      }
-
-      postUpdate();
-
-      return !ignoreSetState;
-    },
-    [apiRef],
-  );
-  return [apiRef.current.state, setGridState, forceUpdate];
-};
+) => [apiRef.current.state, apiRef.current.setState, apiRef.current.forceUpdate] as const;
