+++ packages/grid/_modules_/grid/models/api/gridApi.ts
@@ -1,6 +1,5 @@
 import { GridColumnApi } from './gridColumnApi';
 import { GridColumnMenuApi } from './gridColumnMenuApi';
-import { GridControlStateApi } from './gridControlStateApi';
 import { GridCoreApi } from './gridCoreApi';
 import { GridClipboardApi } from './gridClipboardApi';
 import { GridCsvExportApi } from './gridCsvExportApi';
@@ -52,6 +51,5 @@ export interface GridApi
     GridPrintExportApi,
     GridDisableVirtualizationApi,
     GridLocaleTextApi,
-    GridControlStateApi,
     GridClipboardApi,
     GridScrollApi {}
