+++ packages/grid/_modules_/grid/models/api/gridStateApi.ts
@@ -1,5 +1,6 @@
-import * as React from 'react';
 import { GridState } from '../gridState';
+import {GridControlledStateEventLookup} from "../events";
+import {GridControlStateItem} from "../controlStateItem";
 
 export interface GridStateApi {
   /**
@@ -7,12 +8,21 @@ export interface GridStateApi {
    */
   state: GridState;
   /**
-   * Sets the whole state of the grid.
-   * @param {function} state The new state or a function to return the new state.
-   */
-  setState: (state: GridState | ((previousState: GridState) => GridState)) => void;
-  /**
    * Forces the grid to rerender. It's often used after a state update.
    */
-  forceUpdate: React.Dispatch<any>;
+  forceUpdate: () => void;
+  /**
+   * Sets the whole state of the grid.
+   * @param {GridState | (oldState: GridState) => GridState} state The new state or the callback creating the new state.
+   * @returns {boolean} Has the state been updated.
+   */
+  setState: (state: GridState | ((previousState: GridState) => GridState)) => boolean;
+  /**
+   * Updates a control state that binds the model, the onChange prop, and the grid state together.
+   * @param {GridControlStateItem>} controlState The [[GridControlStateItem]] to be registered.
+   * @ignore - do not document.
+   */
+  unstable_updateControlState: <E extends keyof GridControlledStateEventLookup>(
+      controlState: GridControlStateItem<E>,
+  ) => void;
 }
