+++ packages/grid/_modules_/grid/models/api/index.ts
@@ -19,7 +19,6 @@ export * from './gridColumnMenuApi';
 export * from './gridPreferencesPanelApi';
 export * from './gridPrintExportApi';
 export * from './gridDisableVirtualizationApi';
-export * from './gridControlStateApi';
 export * from './gridClipboardApi';
 export * from './gridCallbackDetails';
 export * from './gridScrollApi';
